Gran avance con el proyecto 	Se agrego la persistencia de la configuracion, se corregierón algunos errores y se agregÂarón otros más...

+++ app/src/main/java/com/autism/figuritas/MyDatabaseApplication.java
@@ -28,7 +28,7 @@ public class MyDatabaseApplication extends Application
         {
             SharedPreferences.Editor editor = sharedPreferences.edit();
 
-            editor.putInt("current_user", 0);
+            editor.putLong("current_user", 0);
             editor.putBoolean("config_complete", false);
             editor.commit();
         }
